+++ students_system.c
@@ -1,371 +1,46 @@
-#include <stdio.h>
-#include <stdlib.h>
-#include <string.h>
-#define N 1001 //Ñ§ÉúÏµÍ³ÈËÊıÉÏÏŞ
-void sort_id();//Ñ§ºÅÅÅĞò
-void sort_rank();//×Ü·Ö³É¼¨ÅÅĞò
-void sort_chi();//ÓïÎÄ³É¼¨ÅÅĞò
-void sort_math();//ÊıÑ§³É¼¨ÅÅĞò
-void sort_eng();//Ó¢Óï³É¼¨ÅÅĞò
-void changestu(int k);//ĞŞ¸ÄÑ§ÉúĞÅÏ¢
-void menu();//ÉùÃ÷²Ëµ¥º¯Êı
-void read();//¶ÁÈ¡´æ´¢Êı¾İÎÄ¼şº¯Êı
-void save();//±£´æÊı¾İÎÄ¼şº¯Êı
-int search_id();//ÉùÃ÷²éÕÒÑ§ºÅº¯Êı
-int search_name();//ÉùÃ÷²éÕÒĞÕÃûº¯Êı
-int search();//ÉùÃ÷²éÕÒ×Üº¯Êı
-
-
-struct Stu {
-	long int id;//Ñ§ºÅ
-	char name[20];//Ñ§ÉúĞÕÃû
-	float chi_mark;//ÓïÎÄ³É¼¨
-	float math_mark;//ÊıÑ§³É¼¨
-	float eng_mark;//Ó¢Óï³É¼¨
-	float sum;//×Ü³É¼¨
-	float ave;//Æ½¾ù·Ö
-} stus[N];
-
-int cnt = 0; //¼ÇÂ¼ÈËÊı
-
-void in_put() { //²åÈëÑ§ÉúĞÅÏ¢º¯Êı
-	while (1) {
-		system("cls");
-		getchar();
-		printf("ÇëÊäÈëÑ§ÉúµÄÑ§ºÅ£º\n");
-		scanf("%ld", &stus[cnt].id);
-		getchar();
-		printf("ÇëÊäÈëÑ§ÉúµÄĞÕÃû£º\n");
-		scanf("%s", stus[cnt].name);
-		getchar();
-		printf("ÇëÊäÈëÑ§ÉúµÄÓïÎÄ³É¼¨£º\n");
-		scanf("%f", &stus[cnt].chi_mark);
-		getchar();
-		printf("ÇëÊäÈëÑ§ÉúµÄÊıÑ§³É¼¨£º\n");
-		scanf("%f", &stus[cnt].math_mark);
-		getchar();
-		printf("ÇëÊäÈëÑ§ÉúµÄÓ¢Óï³É¼¨£º\n");
-		scanf("%f", &stus[cnt].eng_mark);
-		getchar();
-		stus[cnt].sum = stus[cnt].chi_mark + stus[cnt].math_mark + stus[cnt].eng_mark;
-		stus[cnt].ave = (stus[cnt].chi_mark + stus[cnt].math_mark + stus[cnt].eng_mark) / 3;
-		++cnt;
-		printf("ÊÇ·ñ¼ÌĞøÊäÈëÊı¾İ£¿(y/n)\n");
-		char ch;
-		ch=getchar();
-		if (ch == 'y') {
-			continue;
-		}
-		else if (ch == 'n') {
-			break;
-		}
-		else {
-			printf("ÊäÈë´íÎó£¬·µ»ØÖ÷²Ëµ¥\n");
-			system("pause");
-			break;
-		}
-	}
-	
-}
-
-void edit() { //±à¼­Ñ§ÉúĞÅÏ¢º¯Êı
-	char choice;
-	int biao;
-	while (1) {
-		system("cls");
-		getchar();
-		biao=search();
-		if(biao!=N){	
-		printf("ÇëÈ·ÈÏÄãÒªĞŞ¸ÄµÄÑ§ÉúÊÇ·ñÈçÉÏËùÊ¾(y/n)\n");
-				getchar();
-				choice = getchar();
-				if(choice=='y'){
-					changestu(biao);
-				}
-				else if (choice == 'n') {
-					printf("¿ÉÄÜÄãÊä´íÁË£¬ÇëÖØÊÔ\n");
-					system("pause");
-				}
-				else {
-					printf("ÊäÈë´íÎó,ÇëÊäÈë[y/n]\n");
-					system("pause");
-					continue;
-				}
-		}
-		else if(biao==N){
-			continue;
-		}
-		system("cls");
-		printf("ÊÇ·ñ¼ÌĞøĞŞ¸ÄÆäËûÑ§ÉúĞÅÏ¢(y/n):");
-		char ch;
-		getchar();
-		scanf("%c",&ch);
-		if (ch == 'y') {
-			continue;
-		}
-		else {
-			return;
-		}
-	}
-}
-
-void shan() {////É¾³ıÑ§ÉúĞÅÏ¢º¯Êı
-	char mych;
-	int biao;
-	while (1) {
-		system("cls");
-		getchar();
-		biao=search();
-		if(biao!=N){
-			printf("ÊÇ·ñÈ·ÈÏÉ¾³ı[y/n]\n");
-			getchar();
-			mych=getchar();
-			if (mych=='y')
-			{
-				for(int i=biao;i<cnt;++i){
-					stus[biao]=stus[biao+1];
-				}
-				printf("É¾³ı³É¹¦£¡\n");
-				--cnt;
-				system("pause");
-			}
-			else if (mych=='n')
-				{
-				printf("Î´ÄÜÍê³ÉÉ¾³ıÈÎÎñ\n");
-				system("pause");
-			}
-			else {
-				printf("ÊäÈë´íÎó£¬ÇëÖØĞÂ²éÑ¯\n");
-				system("pause");
-				continue;
-			} 
-		}
-		else if(biao==N){
-			continue;
-		}
-		printf("ÊÇ·ñ¼ÌĞøÉ¾³ı(y/n)\n");
-		char ch;
-		getchar();
-		ch=getchar();
-		if (ch == 'y') {
-			continue;
-		}
-		else if (ch == 'n') {
-			break;
-		}
-	}
-}
-
-void view() {//×ÜÀÀÑ§ÉúĞÅÏ¢º¯Êı
-	char mych;
-	system("cls");
-	printf("ÊÇ·ñÆôÓÃ°´×Ü·Ö´Ó¸ßµ½µÍÅÅÃûË³Ğò[y/n]\n");
-	getchar();
-	mych=getchar();
-	if(mych=='y')
-		{
-			system("cls");
-			printf("Ñ§Éú³É¼¨×ÜÀÀ±í(×Ü·ÖÅÅĞò)\n");
-			sort_rank();
-			printf("Ñ§ºÅ\t\tĞÕÃû\t\tÓïÎÄ³É¼¨\tÊıÑ§³É¼¨\tÓ¢Óï³É¼¨\t×Ü·Ö\t\tÆ½¾ù·Ö\n");
-			for (int j = 0; j < cnt; ++j) {
-			printf("%ld\t\t%s\t\t%.2f\t\t%.2f\t\t%.2f\t\t%.2f\t\t%.2f\n", 
-				stus[j].id, stus[j].name, stus[j].chi_mark, stus[j].math_mark, stus[j].eng_mark, stus[j].sum, stus[j].ave);
-			}
-			system("pause");
-		}
-	else if(mych=='n')
-	{
-		sort_id();
-		system("cls");
-		printf("Ñ§Éú³É¼¨×ÜÀÀ±í\n");
-		printf("Ñ§ºÅ\t\tĞÕÃû\t\tÓïÎÄ³É¼¨\tÊıÑ§³É¼¨\tÓ¢Óï³É¼¨\t×Ü·Ö\t\tÆ½¾ù·Ö\n");
-		for (int j = 0; j < cnt; ++j) {
-			printf("%ld\t\t%s\t\t%.2f\t\t%.2f\t\t%.2f\t\t%.2f\t\t%.2f\n", 
-				stus[j].id, stus[j].name, stus[j].chi_mark, stus[j].math_mark, stus[j].eng_mark, stus[j].sum, stus[j].ave);
-		}
-		system("pause");
-	}
-	else
-		{
-			printf("ÊäÈë´íÎó£¬ÇëÖØÊÔ\n");
-			system("pause");
-			view();
-		}
-		
-}
-
-void high() {//¸÷Ñ§¿Æ×î¸ß·Öº¯Êı
-	system("cls");
-	sort_chi();
-	printf("ÓïÎÄ×î¸ß·ÖÑ§Éú£ºÑ§ºÅ:%ld ĞÕÃû:%s ÓïÎÄ³É¼¨:%.2f ÊıÑ§³É¼¨:%.2f Ó¢Óï³É¼¨:%.2f ×Ü·Ö:%.2f Æ½¾ù·Ö:%.2f\n", 
-		stus[0].id, stus[0].name, stus[0].chi_mark, stus[0].math_mark, stus[0].eng_mark, stus[0].sum, stus[0].ave);
-	sort_math();
-	printf("ÊıÑ§×î¸ß·ÖÑ§Éú£ºÑ§ºÅ:%ld ĞÕÃû:%s ÓïÎÄ³É¼¨:%.2f ÊıÑ§³É¼¨:%.2f Ó¢Óï³É¼¨:%.2f ×Ü·Ö:%.2f Æ½¾ù·Ö:%.2f\n", 
-		stus[0].id, stus[0].name, stus[0].chi_mark, stus[0].math_mark, stus[0].eng_mark, stus[0].sum, stus[0].ave);
-	sort_eng();
-	printf("Ó¢Óï×î¸ß·ÖÑ§Éú£ºÑ§ºÅ:%ld ĞÕÃû:%s ÓïÎÄ³É¼¨:%.2f ÊıÑ§³É¼¨:%.2f Ó¢Óï³É¼¨:%.2f ×Ü·Ö:%.2f Æ½¾ù·Ö:%.2f\n", 
-		stus[0].id, stus[0].name, stus[0].chi_mark, stus[0].math_mark, stus[0].eng_mark, stus[0].sum, stus[0].ave);
-	system("pause");
-}
-
-void rank_10() {//×Ü·ÖÇ°Ê®º¯Êı
-	system("cls");
-	sort_rank();
-	printf("Ñ§ºÅ\t\tĞÕÃû\t\tÓïÎÄ³É¼¨\tÊıÑ§³É¼¨\tÓ¢Óï³É¼¨\t×Ü·Ö\t\tÆ½¾ù·Ö\n");
-	for(int j=0;j<10;++j){
-		printf("%ld\t\t%s\t\t%.2f\t\t%.2f\t\t%.2f\t\t%.2f\t\t%.2f\n", 
-			stus[j].id, stus[j].name, stus[j].chi_mark, stus[j].math_mark, stus[j].eng_mark, stus[j].sum, stus[j].ave);
-	}
-	system("pause");
-}
-
-int search() {//²éÕÒ×Üº¯Êı
-	int choice;
-	int biao;
-	while(1){
-		system("cls");
-		printf("ÇëÑ¡Ôñ²éÕÒ·½Ê½[1.Ñ§ºÅ/2.ĞÕÃû](ÓÃÊı×Ö1£¬2½øĞĞÑ¡Ôñ)\n");
-		scanf("%d",&choice);
-		if(choice==1){
-			biao=search_id();
-			break;
-		}
-		else if(choice==2){
-			biao=search_name();
-			break;
-		}
-		else {
-			printf("ÊäÈë´íÎóÇëÖØÊÔ£¡[ÇëÊäÈë1»ò2]\n");
-			system("pause");
-		}
-	}
-	return biao;
-}
-
-void load() { //³õÊ¼»¯Ñ§ÉúÊı¾İº¯Êı
-	char temp[1000]= {0};
-	int j;
-	for (j=0; j < 50; ++j) 
-	{
-		itoa(j+1,temp,10);
-		stus[j].id = 10001 + j;
-		strcpy(stus[j].name,"Ñ§Éú_");
-		strcat(stus[j].name, temp);
-		stus[j].chi_mark = j+49;
-		stus[j].math_mark = j + 50;
-		stus[j].eng_mark = j + 51;
-		stus[j].sum = stus[j].chi_mark + stus[j].math_mark + stus[j].eng_mark;
-		stus[j].ave = (stus[j].chi_mark + stus[j].math_mark + stus[j].eng_mark) / 3;
-		++cnt;
-	}
-	system("pause");
-}
-
-void insert() {//²åÈëº¯Êı
-	int biao;
-	system("cls");
-	sort_id();
-	printf("Ñ§Éú³É¼¨×ÜÀÀ±í\n");
-	printf("[Ñ§ºÅ]\t\tĞÕÃû\t\tÓïÎÄ³É¼¨\tÊıÑ§³É¼¨\tÓ¢Óï³É¼¨\t×Ü·Ö\t\tÆ½¾ù·Ö\n");
-	for (int j = 0; j < cnt; ++j) {
-		printf("%ld\t\t%s\t\t%.2f\t\t%.2f\t\t%.2f\t\t%.2f\t\t%.2f\n", 
-			stus[j].id, stus[j].name, stus[j].chi_mark, stus[j].math_mark, stus[j].eng_mark, stus[j].sum, stus[j].ave);
-	}
-	printf("¿É²åÈëµ½ÊäÈëµÄ[Ñ§ºÅ]Ö®Ç°£¬ÇëÌîĞ´[Ñ§ºÅ]:\n");
-	system("pause");
-	biao=search_id();
-	if(biao!=N){
-		for(int i=cnt;i>=biao;--i){
-			stus[i+1]=stus[i];
-			
-		}
-		printf("ÇëÊäÈë²åÈëÑ§ÉúµÄÑ§ºÅ£º\n");
-		scanf("%ld", &stus[biao].id);
-		getchar();
-		printf("ÇëÊäÈë²åÈëÑ§ÉúµÄĞÕÃû£º\n");
-		scanf("%s", stus[biao].name);
-		getchar();
-		printf("ÇëÊäÈë²åÈëÑ§ÉúµÄÓïÎÄ³É¼¨£º\n");
-		scanf("%f", &stus[biao].chi_mark);
-		getchar();
-		printf("ÇëÊäÈë²åÈëÑ§ÉúµÄÊıÑ§³É¼¨£º\n");
-		scanf("%f", &stus[biao].math_mark);
-		getchar();
-		printf("ÇëÊäÈë²åÈëÑ§ÉúµÄÓ¢Óï³É¼¨£º\n");
-		scanf("%f", &stus[biao].eng_mark);
-		getchar();
-		stus[biao].sum = stus[biao].chi_mark + stus[biao].math_mark + stus[biao].eng_mark;
-		stus[biao].ave = (stus[biao].chi_mark + stus[biao].math_mark + stus[biao].eng_mark) / 3;
-		printf("²åÈë³É¹¦£¡\n");
-		printf("²åÈëµÄÊı¾İ:\n");
-		printf("Ñ§ºÅ\t\tĞÕÃû\t\tÓïÎÄ³É¼¨\tÊıÑ§³É¼¨\tÓ¢Óï³É¼¨\t×Ü·Ö\t\tÆ½¾ù·Ö\n");
-		printf("%ld\t\t%s\t\t%.2f\t\t%.2f\t\t%.2f\t\t%.2f\t\t%.2f\n", 
-			stus[biao].id, stus[biao].name, stus[biao].chi_mark, stus[biao].math_mark, stus[biao].eng_mark, stus[biao].sum, stus[biao].ave);
-		++cnt;
-		system("pause");
-	}
-	else if(biao==N)
-		return;
-	
-}
-
-void faid(){//²»¼°¸ñÃûµ¥º¯Êı
-	system("cls");
-	printf("Ñ§ºÅ\t\tĞÕÃû\t\tÓïÎÄ³É¼¨\tÊıÑ§³É¼¨\tÓ¢Óï³É¼¨\t×Ü·Ö\t\tÆ½¾ù·Ö\n");
-	for(int i=0;i<cnt;++i){
-		if(stus[i].chi_mark<60||stus[i].math_mark<60||stus[i].eng_mark<60){
-			printf("%ld\t\t%s\t\t%.2f\t\t%.2f\t\t%.2f\t\t%.2f\t\t%.2f\n", 
-			stus[i].id, stus[i].name, stus[i].chi_mark, stus[i].math_mark, stus[i].eng_mark, stus[i].sum, stus[i].ave);	
-		}
-	}
-	printf("ÒÔÉÏÎªÓĞ²»¼°¸ñ¿ÆÄ¿µÄÑ§ÉúÃûµ¥\n");
-	system("pause");
-}
-
-
+#include"students_system.h"
 
 int main() {
-	int choose = 1;//Ñ¡Ôñ
+	int choose = 1;//é€‰æ‹©
 	read();
 	do {
 		menu();
 		scanf("%d", &choose);
 		switch (choose) {
-			case 1:
-				in_put();
-				break;
-			case 2:
-				edit();
-				break;
-			case 3:
-				shan();
-				break;
-			case 4:
-				view();
-				break;
-			case 5:
-				high();
-				break;
-			case 6:
-				rank_10();
-				break;
-			case 7:
-				search();
-				break;
-			case 8:
-				load();//³õÊ¼»¯Ñ§ÉúÊı¾İ
-				break;
-			case 9:
-				insert();
-				break;
-			case 10:
-				faid();
-				break;
-			case 0:
-				break;
+		case 1:
+			in_put();
+			break;
+		case 2:
+			edit();
+			break;
+		case 3:
+			shan();
+			break;
+		case 4:
+			view();
+			break;
+		case 5:
+			high();
+			break;
+		case 6:
+			rank_10();
+			break;
+		case 7:
+			search();
+			break;
+		case 8:
+			load();//åˆå§‹åŒ–å­¦ç”Ÿæ•°æ®
+			break;
+		case 9:
+			insert();
+			break;
+		case 10:
+			faid();
+			break;
+		case 0:
+			break;
 			default : {
-				printf("ÊäÈë´íÎó£¡ÇëÖØÊÔ£¡");
+				printf("è¾“å…¥é”™è¯¯ï¼è¯·é‡è¯•ï¼");
 				system("pause");
 			}
 		};
@@ -376,212 +51,4 @@ int main() {
 	printf("\ngood bye~");
 	return 0;
 }
-
-void sort_id() {//Ñ§ºÅÅÅĞò¹¦ÄÜ
-	for (int k = 0; k < cnt - 1; k++) {
-		for (int j = 0; j < cnt - k - 1; ++j) {
-			if (stus[j].id > stus[j + 1].id) {
-				struct Stu rank;
-				rank = stus[j];
-				stus[j] = stus[j + 1];
-				stus[j + 1] = rank;
-			}
-		}
-	}
-}
-
-void sort_rank() {//×Ü³É¼¨ÅÅĞò¹¦ÄÜ
-	for (int k = 0; k < cnt - 1; k++) {
-		for (int j = 0; j < cnt - k - 1; ++j) {
-			if (stus[j].sum < stus[j + 1].sum) {
-				struct Stu rank;
-				rank = stus[j];
-				stus[j] = stus[j + 1];
-				stus[j + 1] = rank;
-			}
-		}
-	}
-}
-
-void sort_chi() {//ÓïÎÄ³É¼¨ÅÅÃû¹¦ÄÜ
-	for (int k = 0; k < cnt - 1; k++) {
-		for (int j = 0; j < cnt - k - 1; ++j) {
-			if (stus[j].chi_mark < stus[j + 1].chi_mark) {
-				struct Stu rank;
-				rank = stus[j];
-				stus[j] = stus[j + 1];
-				stus[j + 1] = rank;
-			}
-		}
-	}
-}
-
-void sort_math() {//ÊıÑ§³É¼¨ÅÅÃû¹¦ÄÜ
-	for (int k = 0; k < cnt - 1; k++) {
-		for (int j = 0; j < cnt - k - 1; ++j) {
-			if (stus[j].math_mark < stus[j + 1].math_mark) {
-				struct Stu rank;
-				rank = stus[j];
-				stus[j] = stus[j + 1];
-				stus[j + 1] = rank;
-			}
-		}
-	}
-}
-
-void sort_eng() {//Ó¢Óï³É¼¨ÅÅÃû¹¦ÄÜ
-	for (int k = 0; k < cnt - 1; k++) {
-		for (int j = 0; j < cnt - k - 1; ++j) {
-			if (stus[j].eng_mark < stus[j + 1].eng_mark) {
-				struct Stu rank;
-				rank = stus[j];
-				stus[j] = stus[j + 1];
-				stus[j + 1] = rank;
-			}
-		}
-	}
-}
-
-
-
-void changestu(int k) {//ĞŞ¸ÄÑ§Éú³É¼¨¹¦ÄÜ
-	int mych;
-while(1){
-	printf("ÇëÑ¡ÔñÄãĞèÒªĞŞ¸ÄµÄ²¿·Ö:\n1.Ñ§ºÅ\n2.ĞÕÃû\n3.ÓïÎÄ³É¼¨\n4.ÊıÑ§³É¼¨\n5.Ó¢Óï³É¼¨\n");
-	scanf("%d",&mych);
-	switch (mych) {
-	case 1:
-		getchar();
-		printf("ÇëÊäÈëÑ§ÉúµÄÑ§ºÅ£º\n");
-		scanf("%ld", &stus[k].id);
-		break;
-	case 2:
-		getchar();
-		printf("ÇëÊäÈëÑ§ÉúµÄĞÕÃû£º\n");
-		scanf("%s", stus[k].name);
-		break;
-	case 3:
-		getchar();
-		printf("ÇëÊäÈëÑ§ÉúµÄÓïÎÄ³É¼¨£º\n");
-		scanf("%f", &stus[k].chi_mark);
-		break;
-	case 4:
-		getchar();
-		printf("ÇëÊäÈëÑ§ÉúµÄÊıÑ§³É¼¨£º\n");
-		scanf("%f", &stus[k].math_mark);
-		break;
-	case 5:
-		getchar();
-		printf("ÇëÊäÈëÑ§ÉúµÄÓ¢Óï³É¼¨£º\n");
-		scanf("%f", &stus[k].eng_mark);
-		break;
-	default:
-		printf("ÊäÈë´íÎó£¬Çë°´ĞòºÅ½øĞĞÑ¡Ôñ£¡\n");
-		system("pause");
-		break;
-	}
-	stus[k].sum = stus[k].chi_mark + stus[k].math_mark + stus[k].eng_mark;
-	stus[k].ave = (stus[k].chi_mark + stus[k].math_mark + stus[k].eng_mark) / 3;
-	printf("ĞŞ¸ÄºóµÄÊı¾İ:\n");
-	printf("Ñ§ºÅ\t\tĞÕÃû\t\tÓïÎÄ³É¼¨\tÊıÑ§³É¼¨\tÓ¢Óï³É¼¨\t×Ü·Ö\t\tÆ½¾ù·Ö\n");
-	printf("%ld\t\t%s\t\t%.2f\t\t%.2f\t\t%.2f\t\t%.2f\t\t%.2f\n", 
-		stus[k].id, stus[k].name, stus[k].chi_mark, stus[k].math_mark, stus[k].eng_mark, stus[k].sum, stus[k].ave);
-	printf("ÊÇ·ñ¼ÌĞøĞŞ¸Ä¸ÃÑ§ÉúµÄĞÅÏ¢(y/n)\n");
-	char mych_1;
-	getchar();
-	scanf("%c",&mych_1);
-	if (mych_1=='y') {
-		continue;
-	}
-	else if (mych_1=='n') {
-		return;
-	}
-	else 
-		{
-		printf("ÊäÈë´íÎó£¬½áÊøĞŞ¸Ä\n");
-		system("pause");
-		return;
-		}
-	}
-}
-
-void menu() {//²Ëµ¥
-	printf("*               ***********\n");
-	printf("*               *\n");
-	printf("*               *\n");
-	printf("*               *\n");
-	printf("*               *\n");
-	printf("********        *********** \n");
-	printf("*      *                  * \n");
-	printf("*      *                  * \n");
-	printf("*      *                  * \n");
-	printf("*      *                  * \n");
-	printf("*      *        *********** \n\n");//logo
-	printf(">>**Ñ§Éú³É¼¨¹ÜÀíÏµÍ³**<< \n");
-	printf("ÇëÑ¡ÔñÏÂÃæµÄÒ»Ïî¹¦ÄÜ£¬ÊäÈë0~10Êı×Ö½øĞĞÑ¡Ôñ£º\n");
-	printf("1.Ñ§Éú³É¼¨Êı¾İÂ¼Èë\n2.Ñ§Éú³É¼¨Êı¾İĞŞ¸Ä\n3.Ñ§Éú³É¼¨Êı¾İÉ¾³ı\n4.Ñ§Éú³É¼¨×ÜÀÀ\n5.²é¿´¸÷Ñ§¿Æ×î¸ß·Ö\n6.²é¿´×Ü·ÖÇ°Ê®\n7.²éÑ¯Ñ§ÉúĞÅÏ¢\n8.³õÊ¼»¯Ñ§ÉúÊı¾İ\n9.²åÈëÑ§ÉúĞÅÏ¢\n10.²éÑ¯²»¼°¸ñÃûµ¥\n0.ÍË³öÏµÍ³\n");}
-
-int search_id() {//²éÕÒÑ§ºÅ¹¦ÄÜ
-	long int f;
-	int j;
-	system("cls");
-	printf("ÇëÊäÈëÄãÏëÒª²éÕÒµÄÑ§ÉúÑ§ºÅ\n");
-	scanf("%ld", &f);
-	for (j = 0; j < cnt; ++j) {
-		if (f==stus[j].id) {
-			printf("Ñ§ºÅ:%ld ĞÕÃû:%s ÓïÎÄ³É¼¨:%.2f ÊıÑ§³É¼¨:%.2f Ó¢Óï³É¼¨:%.2f ×Ü·Ö:%.2f Æ½¾ù·Ö:%.2f\n", stus[j].id, stus[j].name, stus[j].chi_mark, stus[j].math_mark, stus[j].eng_mark, stus[j].sum, stus[j].ave);
-			system("pause");
-			break;
-		} 
-	}
-	if(f!=stus[j].id){
-		printf("Ñ§ºÅ²»´æÔÚ£¬ÇëÖØĞÂÊäÈë\n");
-		system("pause");
-		return N;
-	}
-	return j;
-}
-
-int search_name() {//²éÕÒĞÕÃû¹¦ÄÜ
-	char f[20];
-	int j;
-	system("cls");
-	printf("ÇëÊäÈëÄãÏëÒª²éÕÒµÄÑ§ÉúĞÕÃû\n");
-	getchar();
-	scanf("%s", f);
-	for (j = 0; j < cnt; ++j) {
-		if (strcmp(f,stus[j].name)==0) {
-			printf("Ñ§ºÅ:%ld ĞÕÃû:%s ÓïÎÄ³É¼¨:%.2f ÊıÑ§³É¼¨:%.2f Ó¢Óï³É¼¨:%.2f ×Ü·Ö:%.2f Æ½¾ù·Ö:%.2f\n", stus[j].id, stus[j].name, stus[j].chi_mark, stus[j].math_mark, stus[j].eng_mark, stus[j].sum, stus[j].ave);
-			system("pause");
-			break;
-		} 
-	}
-	if(strcmp(f,stus[j].name)!=0){
-		printf("ĞÕÃû²»´æÔÚ£¬ÇëÖØĞÂÊäÈë\n");
-		system("pause");
-		return N;
-	}
-	return j;
-}
-
-void read()
-{
-	FILE *fp=fopen("data.txt","r+");
-	fscanf(fp,"%d",&cnt);
-	for(int j=0;j<cnt;++j){
-		fscanf(fp,"%ld %s %f %f %f %f %f\n",&stus[j].id, stus[j].name, &stus[j].chi_mark, &stus[j].math_mark, &stus[j].eng_mark, &stus[j].sum, &stus[j].ave);
-	}
-	fclose(fp);
-}
-
-void save()
-{
-	FILE *fp=fopen("data.txt","w+");
-	fprintf(fp,"%d\n",cnt);
-	for(int j=0;j<cnt;++j){
-		fprintf(fp,"%ld %s %.2f %.2f %.2f %.2f %.2f\n",stus[j].id, stus[j].name, stus[j].chi_mark, stus[j].math_mark, stus[j].eng_mark, stus[j].sum, stus[j].ave);
-	}
-	fclose(fp);
-}
-
